add useEffect helper function
