+++ common/helper.ts
@@ -1,7 +1,9 @@
 import moment from 'moment';
-import { Maybe, Message, UpdateEventInput, EventNotificationInput } from '../graphql/generated';
-import { NotificationItem, normaliseEventFormHelperProps } from '../src/types/types';
+import { Maybe, Message, UpdateEventInput, EventNotificationInput, User } from '../graphql/generated';
+import { NotificationItem, normaliseEventFormHelperProps, PeriodType } from '../src/types/types';
 import { periodRate } from '../common/periodTypes';
+import convertMStoTimeLeft from '../common/convertMSToTimeLeft';
+
 
 export const createLink = (
   userEmail: string,
@@ -77,4 +79,115 @@ export const normaliseEventFormHelper = ({eventForm, allDay, files}:  normaliseE
       notifications: normalizedNotifications,
       attachmentIds: files.map(attach => attach.id),
     };
+};
+  
+
+export const handleAllDayHelper = (allDay,
+  setAllDay,
+  dispatch,
+  allDayBufferStartTime,
+  allDayBufferEndTime,
+  eventForm,
+  setAllDayBufferStartTime,
+  setAllDayBufferEndTime) => () => {
+    if (allDay) {
+      setAllDay(false);
+      dispatch({
+        field: 'startTime',
+        value: allDayBufferStartTime,
+      });
+      dispatch({
+        field: 'endTime',
+        value: allDayBufferEndTime,
+      });
+    } else {
+      const isAfter = moment(moment(eventForm.startDate).format('l')).isAfter(
+        eventForm.endDate,
+      );
+
+      if (isAfter) {
+        dispatch({
+          field: 'endDate',
+          value: eventForm.startDate,
+        });
+      }
+
+      setAllDay(true);
+      setAllDayBufferStartTime(eventForm.startTime);
+      setAllDayBufferEndTime(eventForm.endTime);
+
+      dispatch({
+        field: 'startTime',
+        value: '00:00',
+      });
+      dispatch({
+        field: 'endTime',
+        value: '23:59',
+      });
+    }
   };
+
+export const eventDetailsUseEffectHelper = (sharedData,currentUser, event, notificationSettingsData, setNotifications, setSharingUsers ) => {
+      let sharedUsers: User[] = [];
+
+      if (sharedData?.sharedAccess?.targetUsers) {
+        sharedUsers = [...sharedData?.sharedAccess?.targetUsers];
+        const isInclude = sharedUsers.find(({ id }) => id === currentUser.id);
+        if (!isInclude) {
+          sharedUsers.unshift(currentUser);
+        }
+      }
+
+      const initialNotificationPeriod: NotificationItem[] = [];
+
+      const periodTypeMap = {
+        weeks: 'Week',
+        days: 'Day',
+        hours: 'Hour',
+        minutes: 'Minute',
+      };
+
+      if (!event) {
+        if (
+          notificationSettingsData?.notificationSettingsByTag?.items &&
+          notificationSettingsData?.notificationSettingsByTag?.items?.length > 0
+        ) {
+          notificationSettingsData.notificationSettingsByTag.items.forEach(
+            item => {
+              const { type, value } = convertMStoTimeLeft(item.notifyBefore);
+              const tsType = type as keyof typeof periodTypeMap;
+              sharedUsers.forEach(sharedUser => {
+                initialNotificationPeriod.push({
+                  userId: sharedUser.id,
+                  periodType: periodTypeMap[tsType] as PeriodType,
+                  period: value.toString(),
+                });
+              });
+            },
+          );
+        } else {
+          sharedUsers.forEach(sharedUser => {
+            initialNotificationPeriod.push({
+              userId: sharedUser.id,
+              periodType: periodTypeMap.minutes as PeriodType,
+              period: '10',
+            });
+          });
+        }
+      } else if (event?.notifications && event?.notifications?.length > 0) {
+        event.notifications.forEach(notififcation => {
+          const { type, value } = convertMStoTimeLeft(
+            notififcation.notifyBefore,
+          );
+          const tsType = type as keyof typeof periodTypeMap;
+          initialNotificationPeriod.push({
+            userId: notififcation.userId,
+            periodType: periodTypeMap[tsType] as PeriodType,
+            period: value.toString(),
+          });
+        });
+      }
+      setNotifications(initialNotificationPeriod);
+      setSharingUsers(sharedUsers);
+    
+  }
